src/VumbaSoft.AdventureWorks.Controllers/Sales/ContactCreditCards/ContactCreditCardsController.cs
using Microsoft.AspNetCore.Mvc;
using VumbaSoft.AdventureWorks.Objects;
using VumbaSoft.AdventureWorks.Services;
using VumbaSoft.AdventureWorks.Validators;
using System;

namespace VumbaSoft.AdventureWorks.Controllers.Sales
{
    [Area("Sales")]
    public class ContactCreditCardsController : ValidatedController<IContactCreditCardValidator, IContactCreditCardService>
    {
        public ContactCreditCardsController(IContactCreditCardValidator validator, IContactCreditCardService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create(ContactCreditCardView card)
        {
            if (!Validator.CanCreate(card))
                return View(card);

            Service.Create(card);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(Int32 id)
        {
            return NotEmptyView(Service.Get<ContactCreditCardView>(id));
        }

        [HttpGet]
        public ActionResult Edit(Int32 id)
        {
            return NotEmptyView(Service.Get<ContactCreditCardView>(id));
        }

        [HttpPost]
        public ActionResult Edit(ContactCreditCardView card)
        {
            if (!Validator.CanEdit(card))
                return View(card);

            Service.Edit(card);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(Int32 id)
        {
            return NotEmptyView(Service.Get<ContactCreditCardView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(Int32 id)
        {
            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
